识别文件名是否可转换

+++ 识别作品名作者名并新建文件夹移动文件（拖拽，仅多个文件夹，可简繁转换，文件夹名格式[作品][作者]XX）.py
@@ -14,24 +14,24 @@ def main(inputPath):
         if Path.is_dir(Path(aPath)):
             folderName = Path(aPath).name
             nameParts = re.findall("(\\[[^\\]]*\\])", folderName)
-            if fileNameType:
-                workName = OpenCC('t2s').convert(nameParts[0][1:-1])
-                authorName = OpenCC('t2s').convert(nameParts[1][1:-1])
-            else:
-                workName = OpenCC('s2t').convert(nameParts[0][1:-1])
-                authorName = OpenCC('s2t').convert(nameParts[1][1:-1])
-            
-            authorPath = Path(aPath).parent.joinpath(Path(authorName))
-            if not authorPath.exists():
-                Path.mkdir(authorPath)
-            workPath = Path(authorPath).joinpath(Path(workName))
-            if not workPath.exists():
-                Path.mkdir(workPath)
-            for file in Path(aPath).glob('*'):
-                oldFilePath = Path(aPath).joinpath(Path(file.name))
-                newFilePath = Path(workPath).joinpath(Path(file.name))
-                Path(oldFilePath).replace(newFilePath)
-            Path.rmdir(Path(aPath))
+            if len(nameParts) >= 2:
+                if fileNameType:
+                    workName = OpenCC('t2s').convert(nameParts[0][1:-1])
+                    authorName = OpenCC('t2s').convert(nameParts[1][1:-1])
+                else:
+                    workName = OpenCC('s2t').convert(nameParts[0][1:-1])
+                    authorName = OpenCC('s2t').convert(nameParts[1][1:-1])
+                authorPath = Path(aPath).parent.joinpath(Path(authorName))
+                if not authorPath.exists():
+                    Path.mkdir(authorPath)
+                workPath = Path(authorPath).joinpath(Path(workName))
+                if not workPath.exists():
+                    Path.mkdir(workPath)
+                for file in Path(aPath).glob('*'):
+                    oldFilePath = Path(aPath).joinpath(Path(file.name))
+                    newFilePath = Path(workPath).joinpath(Path(file.name))
+                    Path(oldFilePath).replace(newFilePath)
+                Path.rmdir(Path(aPath))
         
 if __name__ == '__main__':
     try:
